hithere v0.0.7: With an 'official' tar line, and the 'guix.scm' is with sha256
'guix build -f guix.scm' builds it into the store
'guix package -f guix.scm' installs this
'hithere' runs it

This uses scripts that exercise the procedures to make a package for the
guix system. The scripts make the hash, of the distcheck target, and
upload to a trial location. Several issues are seen; this is still
a pseudo release, v0.0.7.

Signed-off-by: Gunter Liszewski (s70-10) <gunter.liszewski@gmail.com>

* guix.scm: the local package definition