hithere v0.0: the first release with its developing administrative policy
This is still in the process of finding a proper alignment with agreed and accepted compliance to a release and
development process.

'guix build -f guix.scm' is the command that should build this, here.

Processes will change, and 'guix package -i hithere', with a stable
release, and development policy is the aim, for tis here

Signed-off-by: Gunter Liszewski (s70-10) <gunter.liszewski@gmail.com>

* NEWS: more on this topic of developing policies